+++ src/Foundation/Abstractable/RouteManager.php
@@ -1,5 +1,6 @@
 <?php namespace Orchestra\Foundation\Abstractable;
 
+use Closure;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\NamespacedItemResolver;
 use Orchestra\Extension\RouteGenerator;
@@ -74,17 +75,31 @@ abstract class RouteManager
     /**
      * Return route group dispatch for a package/app.
      *
-     * @param  string   $name   Package name
-     * @return string
+     * @param  string           $name
+     * @param  string           $default
+     * @param  array            $attributes
+     * @param  \Closure|null    $callback
+     * @return array
      */
-    public function group($name, $default, $group = array())
+    public function group($name, $default, $attributes = array(), Closure $callback = null)
     {
         $route = $this->route($name, $default);
 
-        return array_merge($group, array(
+        if ($attributes instanceof Closure) {
+            $callback   = $attributes;
+            $attributes = array();
+        }
+
+        $attributes = array_merge($attributes, array(
             'prefix' => $route->prefix(),
             'domain' => $route->domain(),
         ));
+
+        if (is_callable($callback)) {
+            $this->app['router']->group($attributes, $callback);
+        }
+
+        return $attributes;
     }
 
     /**